Multiple changes to complete the game as per the course instructions...
Adds a score label and implements functionality to increase the score when the bird hits a box.
Gets the score to reset after a game is over and the bird is returned to the original position.
Tidies up the code.
Adds a few TODO items to be tackled someday if time permits.